Finished the rightBar for the fieldTrip
The fieldTrip view should be properly placed in the very last of the 12 columns. Also, fieldTrip places use a full_name attribute instead of a display_name attribute, so I had to add an extra check to determine if display_name or full_name is used. Should close #64, but note that duplicate places appear in the rightBar (since the place was listed multiple times for the fieldTrip).