Added field trip node edges
Close #59

Note that this only adds edges if the fieldtrip is visited AFTER the people/places/stories are added to the graph. Reclicking the fieldtrip from the main view, but not the bottom bar, will correctly update its links.

Similar to the edges for people and places, copy-paste to extract the needed people, places, and stories. Had to use spread syntax in getFieldtripsByID (components/TabViewer/model.js: 122) so that the function doesn't mutate the original fieldtrip data.