Fixed a bug, finished implementing edges
Close #8 , #9 , #80

Turns out I had created issue #80 because some places/people don't have defined people/places/stories (it's empty), so createLinkage() (components/UserNexus/UserNexusModel.js:157) was trying to run .map() on undefined. Fixed that by adding a check to see if those are defined or not.

To add edges for fieldtrips if the person/story/place was clicked after, I just re-ran createLinkage on all fieldtrips once the new node was added to nodeCategories. Also commented the rest of the functions in UserNexusModel