Properly fixed undefined stories
I had incorrectly understood the problem to be "null"s in cstories.json when the problem was actually missing properties in allstories.json.  So I instead moved allstories.json to the /original folder, then created a new program to parse through the JSON, searching for any missing properties and setting them if missing.