Secondary edges
Close issues #99, #100

Basically it just calls the createPrimaryLinkages function to get primary links for the new node, then it calls createPrimaryLInkages again on any primary-linked nodes to find secondary links. Changed arrayTransformation to return an empty array on input of null or undefined, that makes more sense (and made the code a bit simpler).