Handling for bad data in createLinkages()
Close issue #107

I could've simply added an undefined check to createPrimaryLinkages(), but I realized that returning null and giving a console warning in the actual get__ByID() functions in DisplayArtifactModel.js would be a more responsible route (I managed to fix this so quickly because I had a warning saying that placeID "N/A" was not valid). Also, this is sort of a data problem since the placeID was N/A and that caused the function to crash.