No more 100K errors!
Close #106

So, the reason why we were getting errors spewed out so much was because the secondary link implementation was causing the same two nodes to be connected in a bunch of different ways via different connecting nodes; sometimes the same node was being linked back to itself. I believe this has something to do with the numerous links getting the same key for React which it really doesn't like.

Currently, the code filters out the redundant links AFTER they are all generated. I'll probably change this so that the repetitive links are never generated in the first place, instead of having to filter them out later on. This would presumably have the benefit of performance, a little, and more importantly it makes more logical sense to check for duplicate links when creating them.

Also realized that my code was janky and I was giving the wrong attribute for linkNode, so now the proper node name should be used instead of a numeric ID.