Edge filters
Close #10

Added 2 booleans to UserNexus' state to control whether or not primary and secondary links are shown. If false, their respective links are filtered out before the graph is rendered. Also added/edited some comments, and CSS so the 3 buttons look prettier.

Later, as a part of #101, I'll move these buttons to its own tab. For now they will stay on the home screen.