Clicking on a story's keywords leads to a filtered list of stories by the keyword
Close #66

So I had to change a bunch of stuff, but basically when the keywords are clicked it updates the state in TabViewer. This update causes Navigation to be re-rendered using the new keyword, and this keyword is then passed to SearchComponent, which uses it to call handleSearch(), which finally displays the stories that contain the keyword.