Clicking on chapters from the menu mostly takes you to the relevant chapter
Fixes most of #129, but not all of it

Just added a componentWillMount() to BookView to go to the corresponding page in the book when a chapter is clicked. Only implemented for the numbered options in the dropdown (couldn't find the others in the book), and had to hard-code the page numbers. Moved this to a new book_menu.json file, which is also referenced in Heading.js. Added comments, rewrote some of the code that seemed awkward.