Just...coments... and rewriting some code
Close #129

So I just added a bunch of comments. I was trying to figure out how to update the selected chapter from within the view, but apparently that wasn't necessary. Code is still in place for future use if/when we have to dynamically update the page # from outside the BookView