Tabs are draggable!
Close #124

Added a new function for moving a tab at a specific index to a new index. Then I made the tabs draggable, and added a function to detect where tabs are dropped and move the tabs approriately. It's not very pretty or fancy, but the tabs do re-order - I'll probably improve the looks later on. I also ensured that the home tab remains at index 0 (very left), by request of @dhuangg .

Also I removed the console logging in MapView.js because those just spam out the console and make it torturous to see actually relevant logs.