Book tabs partially remember their last page when switched away from
Each book tab's page is no longer stored and changed inside of BookView but has been edited so that the tab's ID (and thus the page it renders) is changed whenever the page to show is changed. For some reason it doesn't save the currently viewed page if you are midway through a chapter on the actual text, only if you go to a new chapter/are looking at the annotation note things at the end of the chapter. I'm guessing it has something to do with how the .epub files are stored and that it counts the main content of a chapter as a single page.