Re-organized tab viewer code a bit
This mostly was about the drag and drop functionality of tabs. I was using the exact same code to calculate the index where the tab would be dropped. This code was used in both handleDrag (callback called while the tab is being dragged) and handleDragEnd (callback called when the tab is released). Moved it to a separate function for simplicity and clarity. No functionality was changed.