Rewriting more code
I poked around a bit in the CSS files this time as well, and found more ways to improve the code. We had quite a few pointless things in the CSS files, and the `active` classes can be rewritten in JS with the use of a boolean expression to control whether or not to render stuff. Also, ETK index + genre + tango indices are now button-y (but don't do anything yet).