Drag and drop functionality for Firefox
Close #160

So Firefox is annoying and makes you set a few things before you can actually drag things, and uses alternate callbacks for drags. I ended up having to store the original and final indices of the drag to the TabViewer class so that I can actually use them in firefox. At least this method is compatible with Safari.