You can actually edit the timeline filters by typing
Close #164

So, we were using the same ref twice for both filters which caused the draggable timelines to override the actual text inputs. Instead, I realized that each change event can be directly used to get the values, so no refs are needed there. Had to change a couple of other things to get it to work, but it should work out properly.