Navigator redux state + reducer
I rewrote + commented more code to simplify the code a bit for when I was moving it to Redux, then I actually did the ugly conversion and migrated stuff. To allow two different Redux action sets to be used together, I used spread syntax to combine them into one actions object. A potential pitfall of this would be if we have identically named actions in both of the actions objects, but so far it works.