Basic Fieldtrip Tool
After an unfortunate number of typos, I finsihed the inchoate fieldtrip tool. Added a new folder for it, as well as a tab type, and added in a map view plus a list of all fieldtrips to select. Because MapView autoupdates with new props, nothing else major was needed. More refinement should follow (my CSS seems to be messing iwth a couple of other things, but I'm not sure). Also removed a couple of console.logs.