Added separate help view pages, rewrote some of the code.
Now, there are buttons to go back and forth between the different, shorter help pages, and the top titles now link to their respective pages. To scroll to top on a page switch, it clears the info which forces the user to be at the top of the page, then loads the new content on the next frame. Also, I rewrote the code to make the top titles simpler to generate and more dynamic.